% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeMLPAparams}
\alias{makeMLPAparams}
\title{Create parameter list for size structured model}
\usage{
makeMLPAparams(m, Linf, k, x0, c, d, xM, xmax, kernel, CR = 4, scale = 100,
  habitat = 1)
}
\arguments{
\item{m}{Instantaneous natural mortality rate}

\item{Linf}{Asymptotic size for von Bertelanffy growth model}

\item{k}{Initial growth rate for von Bertelanffy growth model}

\item{x0}{Initial size for von Bertelanffy growth model (may be negative)}

\item{c}{Coefficent for transforming size into biomass}

\item{d}{Exponent for tranforming size (length) into biomass}

\item{xM}{Age at first reproduction}

\item{xmax}{Maximum age}

\item{kernel}{Collection of square dispersal kernels, arrange in a three-dimensional array with source locations in columns, destination locations in rows, and years along the third dimension of the array}

\item{CR}{The compensation ratio (sets intensity of density dependence; for some reason, it is four for all species!)}

\item{scale}{Equilibrium spawning stock biomass in a deterministic, nonspatial model with unit area}

\item{habitat}{The amount of habitat in each location. May be a scalar (all locations the same) or a vector with length matching the first dimension of kernel}
}
\value{
A list containing all the parameters needed to run MLPAiter
}
\description{
Create parameter list for size structured model
}

